# backend/src/server.ts
import "dotenv/config";
import express from "express";
import bodyParser from "body-parser";
import compression from "compression";
import helmet from "helmet";
import { dbRouter } from "./routes/dbs";
import { tunnelRouter, tunnelTokenExists, resolveAliasHandler } from "./routes/tunnels";
import { adminRouter } from "./routes/admin";
import { meRouter } from "./routes/me";
import { signupRouter } from "./routes/signup";
import { publicRouter } from "./routes/public";
import { appRouter } from "./routes/apps";
import { authMiddleware } from "./middleware/auth";
import { apiRateLimiter, internalAllowTlsRateLimiter } from "./middleware/rate-limit";
import { defaultTimeout } from "./middleware/timeout";
import { ipAllowlist } from "./middleware/ip-allowlist";
import { requestSigning } from "./middleware/request-signing";
import { logger } from "./utils/logger";
import { config } from "./utils/config";
import { makeError } from "./schemas/error";
const ALIAS_REGEX = /^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$/i;
const RESERVED_ALIASES = new Set(["www", "api", "x", "t", "docs", "support", "status", "health", "mail"]);

async function findTokenForAlias(alias: string): Promise<string | null> {
  const { pool } = await import("./db/pool");
  const result = await pool.query(
    `SELECT t.token
     FROM tunnel_aliases a
     JOIN tunnels t ON t.owner_user_id = a.owner_user_id AND t.target_port = a.target_port
     WHERE a.alias = $1 AND t.status = 'active'
     ORDER BY t.created_at DESC
     LIMIT 1`,
    [alias]
  );
  return result.rows?.[0]?.token ?? null;
}

const app = express();

// Security headers
app.use(helmet());

// Response compression (gzip/deflate) - reduces response size by 70-90%
app.use(compression({
  level: 6, // Balanced compression (1-9, higher = smaller but slower)
  threshold: 1024, // Only compress responses > 1KB
  filter: (req, res) => {
    // Don't compress if client doesn't accept it
    if (req.headers["x-no-compression"]) return false;
    return compression.filter(req, res);
  },
}));

// Body parsing with per-route limits (default 10MB, can be overridden per route)
app.use(bodyParser.json({ limit: "10mb" }));

// Request timeout middleware (30 seconds default)
app.use(defaultTimeout);

// Request logging middleware
app.use((req, res, next) => {
  const start = Date.now();
  res.on("finish", () => {
    const duration = Date.now() - start;
    logger.info({
      method: req.method,
      path: req.path,
      status: res.statusCode,
      duration,
      ip: req.ip,
    });
  });
  next();
});

// Health check (no auth, no rate limit)
app.get("/health", (req, res) => {
  res.setHeader("Access-Control-Allow-Origin", "*");
  res.setHeader("Access-Control-Allow-Methods", "GET, OPTIONS");
  res.json({ status: "ok", timestamp: new Date().toISOString() });
});

// Health check endpoints for Kubernetes/Docker
app.get("/health/live", (req, res) => {
  res.json({ status: "alive" });
});

// Cache for ready check (avoid DB query on every health check)
let readyCache: { ready: boolean; checkedAt: number } | null = null;
const READY_CACHE_TTL = 5000; // 5 seconds

app.get("/health/ready", async (req, res) => {
  const now = Date.now();
  
  // Return cached result if fresh
  if (readyCache && (now - readyCache.checkedAt) < READY_CACHE_TTL) {
    if (readyCache.ready) {
      return res.json({ status: "ready", cached: true });
    }
    return res.status(503).json({ status: "not ready", cached: true });
  }
  
  try {
    // Check database connection
    const { pool } = await import("./db/pool");
    await pool.query("SELECT 1");
    readyCache = { ready: true, checkedAt: now };
    res.json({ status: "ready" });
  } catch (error) {
    logger.error({ event: "health_check.failed", error });
    readyCache = { ready: false, checkedAt: now };
    res.status(503).json({ status: "not ready" });
  }
});

// Internal endpoints: IP allowlist + request signing (optional, fail-open if not configured)
// SECURITY: Multiple layers of protection for internal endpoints
app.use("/internal", ipAllowlist(["127.0.0.1", "::1"])); // Default: localhost only
// Note: Request signing is optional (enabled if RELAY_INTERNAL_SECRET is set)
// Caddy may not support custom headers, so we keep query param support for compatibility

// Allow on-demand TLS (Caddy ask endpoint)
// SECURITY: protected by RELAY_INTERNAL_SECRET header + rate-limited + IP allowlist.
app.get("/internal/allow-tls", internalAllowTlsRateLimiter, async (req, res) => {
  try {
    const secret = process.env.RELAY_INTERNAL_SECRET || "";
    if (!secret) {
      // In production we fail at startup if missing; this is a safe fallback.
      return res.status(503).json({ allow: false });
    }
    // Caddy's on-demand TLS `ask` URL can't easily attach custom headers, so we allow either:
    // - header: x-relay-internal-secret
    // - query:  ?secret=...
    // SECURITY: this endpoint is still rate-limited; do not log the secret.
    const providedHeader = req.headers["x-relay-internal-secret"];
    const providedQuery = (req.query.secret as string) || "";
    const provided = providedHeader || providedQuery;
    if (provided !== secret) {
      return res.status(403).json({ allow: false });
    }

    const domain = (req.query.domain as string) || (req.query.host as string) || "";
    const host = domain.split(":")[0].trim().toLowerCase();

    // Token-based hosts: <token>.<TUNNEL_DOMAIN>
    if (host.endsWith(`.${config.tunnelDomain}`)) {
      const token = host.slice(0, -(config.tunnelDomain.length + 1));
      if (!/^[a-zA-Z0-9]{3,64}$/.test(token)) {
        return res.status(403).json({ allow: false });
      }
      const exists = await tunnelTokenExists(token);
      return exists ? res.json({ allow: true }) : res.status(403).json({ allow: false });
    }

    // Alias-based hosts: <alias>.<ALIAS_DOMAIN>
    if (host.endsWith(`.${config.aliasDomain}`)) {
      const alias = host.slice(0, -(config.aliasDomain.length + 1));
      if (!ALIAS_REGEX.test(alias) || RESERVED_ALIASES.has(alias)) {
        return res.status(403).json({ allow: false });
      }
      const token = await findTokenForAlias(alias);
      return token ? res.json({ allow: true }) : res.status(403).json({ allow: false });
    }

    // Everything else: deny
    return res.status(403).json({ allow: false });
  } catch (error) {
    logger.error({ event: "allow_tls.error", error });
    return res.status(500).json({ allow: false });
  }
});

// Internal alias resolution for relay - no auth, uses relay secret
app.get("/internal/resolve-alias", resolveAliasHandler);

// Public endpoints (no auth required) - for website globe visualization
app.use("/public", publicRouter);

// Global rate limiting for all /v1 routes
app.use("/v1", apiRateLimiter);

// Public signup endpoint (no auth required)
app.use("/v1/signup", signupRouter);

// Auth middleware for all other /v1 routes
app.use("/v1", authMiddleware);
app.use("/v1/dbs", dbRouter);
app.use("/v1/tunnels", tunnelRouter);
app.use("/v1/apps", appRouter);
app.use("/v1/admin", adminRouter);
app.use("/v1/me", meRouter);

// Error handling middleware
app.use((err: Error, req: express.Request, res: express.Response, next: express.NextFunction) => {
  logger.error({
    event: "error.unhandled",
    error: err.message,
    stack: err.stack,
    path: req.path,
  });
  res.status(500).json(makeError("INTERNAL_ERROR", "An unexpected error occurred"));
});

// 404 handler
app.use((req, res) => {
  res.status(404).json(makeError("NOT_FOUND", "Route not found"));
});

// Graceful shutdown
const server = app.listen(config.port, () => {
  logger.info({
    event: "server.started",
    port: config.port,
    env: process.env.NODE_ENV || "development",
  });
});

process.on("SIGTERM", () => {
  logger.info({ event: "server.shutdown", signal: "SIGTERM" });
  server.close(() => {
    logger.info({ event: "server.closed" });
    process.exit(0);
  });
});

process.on("SIGINT", () => {
  logger.info({ event: "server.shutdown", signal: "SIGINT" });
  server.close(() => {
    logger.info({ event: "server.closed" });
    process.exit(0);
  });
});